pyproject.toml
[project]
name = "chatbot-v2"
version = "0.1.0"
description = "Agentic RAG Chatbot Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "langchain>=0.1.12",
    "langchain-openai>=0.0.8",
    "langchain-community>=0.0.27",
    "python-dotenv>=1.0.1",
    "chromadb>=0.4.24",
    "tiktoken>=0.6.0",
    "pydantic>=2.6.4",
    "python-multipart>=0.0.9",
    "sentence-transformers>=2.5.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
